+++ src/components/results.js
@@ -9,47 +9,40 @@ import * as signal from "../core/signal-bus.js";
 import { createTestHarnessDisplay } from "./test-harness/layout-harness.js";
 import { gmi } from "../core/gmi/gmi.js";
 
+const getScoreMetaData = result => {
+    if (typeof result === "number") {
+        return { metadata: `SCO=[${result}]` };
+    }
+    if (typeof result === "string") {
+        const digitsRegex = /\d+/;
+        const score = result.match(digitsRegex);
+        return score ? { metadata: `SCO=[${score}]` } : undefined;
+    }
+};
+
+const fireGameCompleteStat = result => {
+    gmi.sendStatsEvent("score", "display", getScoreMetaData(result));
+};
+
 export class Results extends Screen {
     constructor() {
         super();
     }
 
-    getScoreMetaData(result) {
-        if (typeof result === "number") {
-            return { metadata: `SCO=[${result}]` };
-        }
-        if (typeof result === "string") {
-            const digitsRegex = /\d+/;
-            const score = result.match(digitsRegex);
-            return score ? { metadata: `SCO=[${score}]` } : undefined;
-        }
-    }
-
-    fireGameCompleteStat(result) {
-        gmi.sendStatsEvent("score", "display", this.getScoreMetaData(result));
-    }
-
     create() {
         const theme = this.context.config.theme[this.game.state.current];
-
-        const backgroundImage = this.game.add.image(0, 0, "results.background");
-        this.scene.addToBackground(backgroundImage);
-
-        const titleImage = this.scene.addToBackground(this.game.add.image(0, -150, "results.title"));
-        this.scene.addToBackground(titleImage);
+        this.scene.addToBackground(this.game.add.image(0, 0, "results.background"));
+        this.scene.addToBackground(this.scene.addToBackground(this.game.add.image(0, -150, "results.title")));
 
         const resultsText = this.game.add.text(0, 50, this.transientData.results, theme.resultText.style);
         this.scene.addToBackground(resultsText);
 
-        // TODO the following line should be added back once the overlap issue is resolved NT:04:02:19
-        const showAchievements = false; //!!this.context.config.theme.game.achievements;
-        const defaultButtons = ["pause", "restart", "continueGame"];
-        const buttons = showAchievements ? defaultButtons.concat("achievements") : defaultButtons;
-
-        this.scene.addLayout(buttons);
+        const achievements = this.context.config.theme.game.achievements ? ["achievements"] : [];
+        const buttons = ["pause", "restart", "continueGame"];
+        this.scene.addLayout(buttons.concat(achievements));
         createTestHarnessDisplay(this.game, this.context, this.scene);
 
-        this.fireGameCompleteStat(this.transientData.results);
+        fireGameCompleteStat(this.transientData.results);
 
         signal.bus.subscribe({
             name: "continue",