+++ src/core/layout/group.js
@@ -6,7 +6,6 @@
 import fp from "../../../lib/lodash/fp/fp.js";
 
 import * as ButtonFactory from "./button-factory.js";
-import { applyButtonOverrides } from "./button-overrides.js";
 
 const horizontal = {
     left: (metrics, group, horizontalsType) => {
@@ -105,7 +104,11 @@ export class Group extends Phaser.Group {
         const invScale = 1 / metrics.scale;
         this.scale.setTo(invScale, invScale);
         this._setGroupPosition(metrics);
-        applyButtonOverrides(metrics.scale, this._buttons);
+
+        this._buttons.forEach(button => {
+            button.x = button.x + button.shiftX * metrics.scale;
+            button.y = button.y + button.shiftY * metrics.scale;
+        });
     }
 
     alignChildren() {