+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 | Version | Description |
 |---------|-------------|
+| 2.0.0 | |
+| | Achievements stats amendments. |
+| | Add achievement button config for results screen. |
+| | Add per screen button overrides. |
 | | Update webpack config so bowser package is run through babel. |
 | | Add achievement notification to achievement button. |
 | 1.0.14 | |